update implementation and specs
