+++ packages/jsonapi-to-model/tsconfig.json
@@ -9,8 +9,8 @@
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
     "useDefineForClassFields": true,
+    "resolveJsonModule": true,
   },
   "files": [],
   "include": [],
